requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
lpips>=0.1.4
imageio>=2.9.0

SQLAlchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0

PyYAML>=5.4.0
python-dotenv>=0.19.0

argparse>=1.4.0
rich>=10.0.0
tqdm>=4.62.0

Flask>=2.0.0

pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

